+++ my-app/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import { BrowserRouter as Router, Route } from "react-router-dom";
-import { Switch, Link } from 'react-router';
 import './App.css'
+import { BrowserRouter as Router, Route } from "react-router-dom";
+import { Switch } from 'react-router';
 import Home from './components/Home'
 import Login from './components/Login'
 import Register from './components/Register'
@@ -10,7 +10,7 @@ function App() {
   return (
     <Router>
       <Switch> 
-        <Route path='/' component={Home}/>
+        <Route exact path='/' component={Home} />
         <Route path='/Login' component={Login}/>
         <Route path='/Register' component={Register}/>
       </Switch>
