--- my-app/src/index.css
@@ -1,9 +1,0 @@
-* {
-  text-align: center;
-  background-color: #FEDBD0 ;
-}
-/*
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
-    monospace;
-} */
