--- notes-for-you-app/src/components/Home.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import { Link } from "react-router-dom";
-
-const Home = () => (
-  <div >
-  <Link to = '/Login'><button>Login</button></Link>
-  <button>SignUp</button>
-  </div>
-)
-
-export default Home
