--- notes-for-you-app/src/components/Register.js
@@ -1,17 +1,0 @@
-import React from 'react'
-
-const Register = () => {
-    return (
-      <>
-        <form id="register">
-          <input placeholder="Email"></input>
-          <input placeholder="Password"></input>
-          <input placeholder="Confirm your password"></input>
-          <button>Sign Up</button>
-          <button>Sign Up with Google</button>
-        </form>
-      </>
-    );
-  };
-  
-  export default Register;
