--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "react-router-dom": "^5.3.0"
-  }
-}
